impl(WIP): handlePlayingSong and seekbar

+++ app/src/main/java/vip/yazilim/p2g/android/util/helper/TimeHelper.kt
@@ -15,7 +15,7 @@ class TimeHelper {
             DateTimeFormatter.ofPattern("dd.MM.YYYY HH:mm")
 
         @SuppressLint("SimpleDateFormat")
-        fun getHumanReadableTimestamp(time: Long): String {
+        fun getHumanReadableTimestamp(time: Int): String {
             return (SimpleDateFormat("mm:ss")).format(time)
         }
     }
